page 404 login arraanco pongS
y

Alright - we've been working for a while now, and we're 3 or so cocktails into an engineer's Saturday night, but we've got some work left to do: Creating our actual game object. Our game object holds references to the other objects in the game (paddles and ball in this case), and handles all of our logical calculations, including collisions and game over checking. This file's a little long, so lets take a look at the stub of our class to explain this all a little better: